refactor: HttpRequest의 일급콜렉션화 리팩토링
