+++ src/main/java/webserver/controller/user/UserCreateController.java
@@ -4,7 +4,7 @@ import db.DataBase;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.http.HttpRequest;
+import webserver.http.request.HttpRequest;
 import webserver.http.HttpResponse;
 import webserver.controller.RequestController;
